client/pyproject.toml
[tool.poetry]
name = "edna-client"
version = "0.1.0"
description = ""
authors = ["Cyber-Physical Systems Group, Technical University of Munich <commonroad@lists.lrz.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
typer = "^0.12.3"
fastapi = "^0.111.0"
python-jose = "^3.3.0"
yarl = "^1.9.4"
logging = "^0.4.9.6"
keyring = "^25.2.1"
tinynetrc = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
